requirements.txt
numpy==1.21.2
shapely==1.7.0
pygame==2.1.2
gym==0.18.0
heapdict==1.0.1
opencv-python==4.5.4.58
scipy==1.7.1
tensorboard==2.8.0
tqdm==4.62.2
matplotlib==3.4.3
einops==0.6.1
pytorch==1.9.1
